refactor: rename to WikiMind, add Clip/Ask tab bar
- Rename extension from "WikiMind Clipper" to "WikiMind"
- Add two-tab layout (Clip / Ask) to the popup
- Extract clip logic into ClipTab component
- Add AskTab placeholder ("coming soon")
- Add browser_specific_settings.gecko.id for Firefox AMO
