+++ apps/web-extension/src/popup/components/AskTab.tsx
@@ -1,0 +1,19 @@
+export function AskTab() {
+  return (
+    <div
+      style={{
+        textAlign: "center",
+        padding: "24px 16px",
+        color: "#94a3b8",
+      }}
+    >
+      <div style={{ fontSize: "32px", marginBottom: "8px" }}>&#128269;</div>
+      <p style={{ fontSize: "14px", fontWeight: 600, margin: "0 0 4px" }}>
+        Ask your wiki
+      </p>
+      <p style={{ fontSize: "12px", margin: 0 }}>
+        Search and query your knowledge base — coming soon.
+      </p>
+    </div>
+  );
+}
